Quick wins: future date filter, float32 GPU, vectorized fair multiples, tensor_ops, worker limit
1. backfill_edgar: drop future as_of_date rows before merge (EDGAR returns
   estimated future quarters); tested — future tickers dropped, past kept.

2. fractal_windows_gpu: tensors use float32 (2x faster on MX550 vs float64).

3. damodaran_data: compute_fair_multiples vectorized (removed iterrows loop).

4. tensor_ops.py (new): unified GPU/CPU rolling ops — sum, mean, std, slope, beta.
   Used by rolling_window_analysis.rolling_cumsum_2d and peer_analytics.

5. peer_analytics: analyze_fundamental_trends builds wide [tickers, quarters]
   panel per metric; uses tensor_ops.rolling_slope on GPU when available.

6. run_daily_automation already had --max-workers (default 4) with
   ThreadPoolExecutor; verified working.

+++ fractal_windows_gpu.py
@@ -74,8 +74,8 @@ def fractal_batch(wide_logp: np.ndarray, a: int = 30, b: int = 3,
     spans = spans_generator(a, b)
     lengths = sorted({t - f for f, t in spans})
 
-    logp = torch.as_tensor(wide_logp, dtype=torch.float64, device=dev)
-    idx = torch.arange(D, dtype=torch.float64, device=dev)
+    logp = torch.as_tensor(wide_logp, dtype=torch.float32, device=dev)
+    idx = torch.arange(D, dtype=torch.float32, device=dev)
     k_y = idx * logp  # [T, D]  k * logp
     dr = torch.diff(logp, dim=1, prepend=logp[:, :1] * float("nan"))  # daily ret
